refactor: moved public C API components in the core crate to the flux crate
This puts all of the C api in one place.
