+++ libflux/src/core/lib.rs
@@ -14,26 +14,16 @@ pub mod scanner;
 pub mod semantic;
 
 use std::error;
-use std::ffi::*;
 use std::fmt;
-use std::os::raw::c_char;
-
-use parser::Parser;
 
 pub use ast::DEFAULT_PACKAGE_NAME;
-
-/// An error handle designed to allow passing `Error` instances to library
-/// consumers across language boundaries.
-pub struct ErrorHandle {
-    /// A heap-allocated `Error`
-    pub err: Box<dyn error::Error>,
-}
 
 /// An error that can occur due to problems in ast generation or semantic
 /// analysis.
 #[derive(Debug, Clone)]
 pub struct Error {
-    msg: String,
+    /// Message.
+    pub msg: String,
 }
 
 impl fmt::Display for Error {
@@ -73,409 +63,5 @@ impl From<semantic::check::Error> for Error {
         Error {
             msg: format!("{}", err),
         }
-    }
-}
-
-/// Frees a previously allocated error.
-///
-/// Note: we use the pattern described here: https://doc.rust-lang.org/std/boxed/index.html#memory-layout
-/// wherein a pointer where ownership is being transferred is modeled as a Box, and if it could be
-/// null, then it's wrapped in an Option.
-#[no_mangle]
-pub extern "C" fn flux_free_error(_err: Option<Box<ErrorHandle>>) {}
-
-/// Frees a pointer to characters.
-///
-/// # Safety
-///
-/// This function is unsafe because improper use may lead to
-/// memory problems. For example, a double-free may occur if the
-/// function is called twice on the same raw pointer.
-#[no_mangle]
-pub unsafe extern "C" fn flux_free_bytes(cstr: *mut c_char) {
-    Box::from_raw(cstr);
-}
-
-/// A buffer of flux source.
-#[repr(C)]
-pub struct flux_buffer_t {
-    /// A pointer to a byte array.
-    pub data: *const u8,
-    /// The length of the byte array.
-    pub len: usize,
-}
-
-/// flux_parse parses a string containing Flux source code into an AST.
-///
-/// # Safety
-///
-/// This function is unsafe because it dereferences a raw pointer passed
-/// in as a parameter. For example, if that pointer is NULL, undefined behavior
-/// could occur.
-#[no_mangle]
-pub unsafe extern "C" fn flux_parse(
-    cfname: *const c_char,
-    csrc: *const c_char,
-) -> Box<ast::Package> {
-    let fname = String::from_utf8(CStr::from_ptr(cfname).to_bytes().to_vec()).unwrap();
-    let src = String::from_utf8(CStr::from_ptr(csrc).to_bytes().to_vec()).unwrap();
-    let mut p = Parser::new(&src);
-    let pkg: ast::Package = p.parse_file(fname).into();
-    Box::new(pkg)
-}
-
-/// flux_ast_get_error returns the first error in the given AST.
-///
-/// # Safety
-///
-/// This funtion is unsafe because it dereferences a raw pointer.
-#[no_mangle]
-pub unsafe extern "C" fn flux_ast_get_error(
-    ast_pkg: *const ast::Package,
-) -> Option<Box<ErrorHandle>> {
-    let ast_pkg = ast::walk::Node::Package(&*ast_pkg);
-    let mut errs = ast::check::check(ast_pkg);
-    if !errs.is_empty() {
-        let err = Vec::remove(&mut errs, 0);
-        Some(Box::new(ErrorHandle { err: Box::new(err) }))
-    } else {
-        None
-    }
-}
-
-/// Frees an AST package.
-///
-/// Note: we use the pattern described here: https://doc.rust-lang.org/std/boxed/index.html#memory-layout
-/// wherein a pointer where ownership is being transferred is modeled as a Box, and if it could be
-/// null, then it's wrapped in an Option.
-#[no_mangle]
-pub extern "C" fn flux_free_ast_pkg(_: Option<Box<ast::Package>>) {}
-
-/// # Safety
-///
-/// This function is unsafe because it dereferences a raw pointer passed
-/// in as a parameter. For example, if that pointer is NULL, undefined behavior
-/// could occur.
-#[no_mangle]
-pub unsafe extern "C" fn flux_parse_json(
-    cstr: *mut c_char,
-    out_pkg: *mut Option<Box<ast::Package>>,
-) -> Option<Box<ErrorHandle>> {
-    let buf = CStr::from_ptr(cstr).to_bytes(); // Unsafe
-    let res: Result<ast::Package, serde_json::error::Error> = serde_json::from_slice(buf);
-    match res {
-        Ok(pkg) => {
-            *out_pkg = Some(Box::new(pkg));
-            None
-        }
-        Err(err) => {
-            let errh = ErrorHandle { err: Box::new(err) };
-            Some(Box::new(errh))
-        }
-    }
-}
-
-/// # Safety
-///
-/// This function is unsafe because it dereferences raw pointers passed
-/// in as parameters. For example, if that pointer is NULL, undefined behavior
-/// could occur.
-#[no_mangle]
-pub unsafe extern "C" fn flux_ast_marshal_json(
-    ast_pkg: *const ast::Package,
-    buf: *mut flux_buffer_t,
-) -> Option<Box<ErrorHandle>> {
-    let ast_pkg = &*ast_pkg;
-    let data = match serde_json::to_vec(ast_pkg) {
-        Ok(v) => v,
-        Err(err) => {
-            let errh = ErrorHandle { err: Box::new(err) };
-            return Some(Box::new(errh));
-        }
-    };
-
-    (*buf).len = data.len();
-    (*buf).data = Box::into_raw(data.into_boxed_slice()) as *mut u8;
-    None
-}
-
-/// flux_ast_marshal_fb serializes the given AST package to a flatbuffer.
-///
-/// # Safety
-///
-/// This function is unsafe because it takes a dereferences a raw pointer passed
-/// in as a parameter. For example, if that pointer is NULL, undefined behavior
-/// could occur.
-#[no_mangle]
-pub unsafe extern "C" fn flux_ast_marshal_fb(
-    ast_pkg: *const ast::Package,
-    buf: *mut flux_buffer_t,
-) -> Option<Box<ErrorHandle>> {
-    let ast_pkg = &*ast_pkg;
-    let (mut vec, offset) = match ast::flatbuffers::serialize(ast_pkg) {
-        Ok(vec_offset) => vec_offset,
-        Err(err) => {
-            let errh = ErrorHandle {
-                err: Box::new(Error::from(err)),
-            };
-            return Some(Box::new(errh));
-        }
-    };
-
-    // Note, split_off() does a copy: https://github.com/influxdata/flux/issues/2194
-    let data = vec.split_off(offset);
-    (*buf).len = data.len();
-    (*buf).data = Box::into_raw(data.into_boxed_slice()) as *mut u8;
-    None
-}
-
-/// Frees a semantic package.
-#[no_mangle]
-pub extern "C" fn flux_free_semantic_pkg(_: Option<Box<semantic::nodes::Package>>) {}
-
-/// flux_semantic_marshal_fb populates the supplied buffer with a FlatBuffers serialization
-/// of the given AST.
-///
-/// # Safety
-///
-/// This function is unsafe because it takes a dereferences a raw pointer passed
-/// in as a parameter. For example, if that pointer is NULL, undefined behavior
-/// could occur.
-#[no_mangle]
-pub unsafe extern "C" fn flux_semantic_marshal_fb(
-    sem_pkg: *const semantic::nodes::Package,
-    buf: *mut flux_buffer_t,
-) -> Option<Box<ErrorHandle>> {
-    let sem_pkg = &*sem_pkg;
-    let (mut vec, offset) = match semantic::flatbuffers::serialize(sem_pkg) {
-        Ok(vec_offset) => vec_offset,
-        Err(err) => {
-            let errh = ErrorHandle {
-                err: Box::new(Error::from(err)),
-            };
-            return Some(Box::new(errh));
-        }
-    };
-
-    // Note, split_off() does a copy: https://github.com/influxdata/flux/issues/2194
-    let data = vec.split_off(offset);
-    (*buf).len = data.len();
-    (*buf).data = Box::into_raw(data.into_boxed_slice()) as *mut u8;
-    None
-}
-
-/// flux_error_str returns the error message associated with the given error.
-///
-/// # Safety
-///
-/// This function is unsafe because it dereferences a raw pointer passed as a
-/// parameter
-#[no_mangle]
-pub unsafe extern "C" fn flux_error_str(errh: *const ErrorHandle) -> CString {
-    let errh = &*errh;
-    CString::new(format!("{}", errh.err)).unwrap()
-}
-
-/// # Safety
-///
-/// This function is unsafe because it dereferences a raw pointer passed as a
-/// parameter
-///
-/// flux_merge_ast_pkg_files merges the files of a given input ast::Package into the file
-/// vector of an output ast::Package.
-#[no_mangle]
-pub unsafe extern "C" fn flux_merge_ast_pkgs(
-    out_pkg: *mut ast::Package,
-    in_pkg: *mut ast::Package,
-) -> Option<Box<ErrorHandle>> {
-    // Do not change ownership here so that Go maintains ownership of packages
-    let out_pkg = &mut *out_pkg;
-    let in_pkg = &mut *in_pkg;
-
-    match merge_packages(out_pkg, in_pkg) {
-        None => None,
-        Some(err) => {
-            let err_handle = ErrorHandle { err: Box::new(err) };
-            Some(Box::new(err_handle))
-        }
-    }
-}
-
-/// merge_packages takes an input package and an output package, checks that the package
-/// clauses match and merges the files from the input package into the output package. If
-/// package clauses fail validation then an option with an Error is returned.
-pub fn merge_packages(out_pkg: &mut ast::Package, in_pkg: &mut ast::Package) -> Option<Error> {
-    let out_pkg_clause = match &out_pkg.files[0].package {
-        Some(clause) => Some(&clause.name.name),
-        None => None,
-    };
-
-    // Check that all input files have a package clause that matches the output package.
-    for file in &in_pkg.files {
-        let file_clause = match &file.package {
-            Some(clause) => Some(&clause.name.name),
-            None => None,
-        };
-
-        match (out_pkg_clause, file_clause) {
-            (Some(out_name), Some(in_name)) => {
-                if out_name != in_name {
-                    return Some(Error::from(format!(
-                        r#"file's package clause: "{}" does not match package output package clause: "{}""#,
-                        in_name, out_name
-                    )));
-                }
-            }
-            (None, Some(in_name)) => {
-                return Some(Error::from(format!(
-                    r#"output package does not have a package clause, but current file has package clause "{}""#,
-                    in_name
-                )))
-            }
-            (Some(out_name), None) => {
-                return Some(Error::from(format!(
-                    r#"current file does not have a package clause, but output package has package clause "{}""#,
-                    out_name
-                )))
-            }
-            (None, None) => (),
-        };
-    }
-    out_pkg.files.append(&mut in_pkg.files);
-    None
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{ast, flux_ast_get_error, merge_packages};
-
-    #[test]
-    fn ok_merge_multi_file() {
-        let in_script = "package foo\na = 1\n";
-        let out_script = "package foo\nb = 2\n";
-
-        let in_file = crate::parser::parse_string("test", in_script);
-        let out_file = crate::parser::parse_string("test", out_script);
-        let mut in_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![in_file.clone()],
-        };
-        let mut out_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![out_file.clone()],
-        };
-        merge_packages(&mut out_pkg, &mut in_pkg);
-        let got = out_pkg.files;
-        let want = vec![out_file, in_file];
-        assert_eq!(want, got);
-    }
-
-    #[test]
-    fn ok_no_in_pkg() {
-        let out_script = "package foo\nb = 2\n";
-
-        let out_file = crate::parser::parse_string("test", out_script);
-        let mut in_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![],
-        };
-        let mut out_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![out_file.clone()],
-        };
-        merge_packages(&mut out_pkg, &mut in_pkg);
-        let got = out_pkg.files;
-        let want = vec![out_file];
-        assert_eq!(want, got);
-    }
-
-    #[test]
-    fn err_no_out_pkg_clause() {
-        let in_script = "package foo\na = 1\n";
-        let out_script = "";
-
-        let in_file = crate::parser::parse_string("test", in_script);
-        let out_file = crate::parser::parse_string("test", out_script);
-        let mut in_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![in_file.clone()],
-        };
-        let mut out_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![out_file.clone()],
-        };
-        let got_err = merge_packages(&mut out_pkg, &mut in_pkg).unwrap().msg;
-        let want_err = r#"output package does not have a package clause, but current file has package clause "foo""#;
-        assert_eq!(got_err.to_string(), want_err);
-    }
-
-    #[test]
-    fn err_no_in_pkg_clause() {
-        let in_script = "";
-        let out_script = "package foo\nb = 100\n";
-
-        let in_file = crate::parser::parse_string("test", in_script);
-        let out_file = crate::parser::parse_string("test", out_script);
-        let mut in_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![in_file.clone()],
-        };
-        let mut out_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![out_file.clone()],
-        };
-        let got_err = merge_packages(&mut out_pkg, &mut in_pkg).unwrap().msg;
-        let want_err = r#"current file does not have a package clause, but output package has package clause "foo""#;
-        assert_eq!(got_err.to_string(), want_err);
-    }
-
-    #[test]
-    fn ok_no_pkg_clauses() {
-        let in_script = "a = 100\n";
-        let out_script = "b = a * a\n";
-        let in_file = crate::parser::parse_string("test", in_script);
-        let out_file = crate::parser::parse_string("test", out_script);
-        let mut in_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![in_file.clone()],
-        };
-        let mut out_pkg = ast::Package {
-            base: Default::default(),
-            path: "./test".to_string(),
-            package: "foo".to_string(),
-            files: vec![out_file.clone()],
-        };
-        let result = merge_packages(&mut out_pkg, &mut in_pkg);
-        assert!(result.is_none());
-        assert_eq!(2, out_pkg.files.len());
-    }
-
-    #[test]
-    fn test_ast_get_error() {
-        let ast = crate::parser::parse_string("test", "x = 3 + / 10 - \"");
-        let ast = Box::into_raw(Box::new(ast.into()));
-        let errh = unsafe { flux_ast_get_error(ast) };
-        assert_eq!(
-            "error at test@1:9-1:10: invalid expression: invalid token for primary expression: DIV",
-            format!("{}", errh.unwrap().err)
-        );
     }
 }
