+++ libflux/src/flux/Cargo.toml
@@ -12,6 +12,10 @@ crate-type = ["rlib", "staticlib", "cdylib"]
 [dependencies]
 core = { path = "../core" }
 flatbuffers = "0.6.0"
+serde = "^1.0.59"
+serde_derive = "^1.0.59"
+serde_json = "1.0"
+serde-aux = "0.6.1"
 
 [dev-dependencies]
 maplit = "1.0.2"
